requirements.txt
pydantic[dotenv]
mypy
pytest-cov
PyYAML